# openapi.yaml
openapi: 3.0.3
info:
  title: microserver
  description: >
    Servidor HTTP en Go sobre SQLite (sqlite-vec) con embeddinggemma (Ollama)
    para embeddings: búsqueda semántica, colecciones dinámicas con JSON libre,
    filtrado/orden/agregación/full-text, y multi-usuario con roles.
    Ver README.md para instalación, despliegue y detalles de arquitectura.
  version: "0.1.0"
  license:
    name: MIT
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
...
paths:
...
  /login:
    post:
      tags: [auth]
      summary: Login
      operationId: login
      description: >
        Autentica contra la tabla de usuarios y devuelve un token firmado
        (HMAC) que incluye el rol del usuario, válido por 1 hora.
        Sujeto a rate limit por IP (5 intentos/min).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Login exitoso
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResponse" }
        "401":
          description: Credenciales inválidas
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429":
          description: Demasiados intentos
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /admin/backup:
    post:
      tags: [admin]
      summary: Disparar un backup inmediato
      operationId: triggerBackup
      description: Requiere rol admin. Mismo VACUUM INTO + rotación que el backup automático cada 1h.
      responses:
        "201":
          description: Backup creado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BackupInfo" }
        "403":
          description: No es admin
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /admin/backups:
    get:
      tags: [admin]
      summary: Listar backups
      operationId: listBackups
      description: Requiere rol admin. Más nuevo primero.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/BackupInfo" }
        "403":
          description: No es admin
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /admin/backups/{name}:
    get:
      tags: [admin]
      summary: Descargar un backup
      operationId: downloadBackup
      description: >
        Requiere rol admin. name se valida contra el patrón exacto que
        genera el servidor — cualquier otra cosa (incluidos intentos de
        path traversal) da 400 antes de tocar el filesystem. No hay
        endpoint de restore: reemplazar la base de datos en caliente no es
        seguro sin un cambio de arquitectura mayor — es un procedimiento
        manual offline (ver README).
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string, pattern: '^backup-\d{8}-\d{6}\.\d{9}\.db$' }
      responses:
        "200":
          description: Contenido del backup
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        "400":
          description: Nombre inválido
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: No es admin
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: No existe
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /users/me/password:
    put:
      tags: [auth]
      summary: Cambiar la propia contraseña
      operationId: changePassword
      description: Cualquier usuario autenticado. El usuario objetivo sale del token, nunca de la URL.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChangePasswordRequest" }
      responses:
        "204":
          description: Contraseña cambiada
        "400":
          description: Contraseña nueva demasiado corta (mínimo 8 caracteres)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: current_password incorrecta, o token ausente/inválido/expirado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /users:
    post:
      tags: [auth]
      summary: Crear usuario
      operationId: createUser
      description: Requiere rol admin.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateUserRequest" }
      responses:
        "201":
          description: Usuario creado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserInfo" }
        "400":
          description: Username/rol inválido, o contraseña muy corta
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: No es admin
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: El usuario ya existe
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    get:
      tags: [auth]
      summary: Listar usuarios
      operationId: listUsers
      description: Requiere rol admin. Nunca incluye el hash de contraseña.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/UserInfo" }
        "403":
          description: No es admin
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /users/{username}:
    delete:
      tags: [auth]
      summary: Borrar usuario
      operationId: deleteUser
      description: Requiere rol admin. Rechaza borrar el último admin que queda.
      parameters:
        - $ref: "#/components/parameters/Username"
      responses:
        "204":
          description: Usuario borrado
        "403":
          description: No es admin
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Usuario no encontrado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: Es el último admin, no se puede borrar
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /users/{username}/password:
    put:
      tags: [auth]
      summary: Resetear la contraseña de otro usuario
      operationId: resetPassword
      description: Requiere rol admin. No necesita conocer la contraseña actual — salida para cuentas bloqueadas.
      parameters:
        - $ref: "#/components/parameters/Username"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ResetPasswordRequest" }
      responses:
        "204":
          description: Contraseña reseteada
        "400":
          description: Contraseña nueva demasiado corta (mínimo 8 caracteres)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: No es admin
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Usuario no encontrado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /items:
    get:
      tags: [items]
      summary: Listar items
      operationId: listItems
      description: Cualquier usuario autenticado.
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Item" }
    post:
      tags: [items]
      summary: Insertar un item
      operationId: insertItem
      description: Requiere rol admin. Genera el embedding (embeddinggemma, 768 dims), guardado en float32 + binario cuantizado.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InsertRequest" }
      responses:
        "201":
          description: Creado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InsertResponse" }
        "400":
          description: Falta text, o JSON inválido
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: No es admin
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /items/bulk:
    post:
      tags: [items]
      summary: Insertar varios items
      operationId: bulkInsertItems
      description: >
        Requiere rol admin. Todo-o-nada: si un item falla, no se inserta
        ninguno del lote. Máximo 100 items por request.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items: { $ref: "#/components/schemas/InsertRequest" }
      responses:
        "201":
          description: Creados, mismo orden que el request
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/InsertResponse" }
        "400":
          description: Lote vacío, supera 100 items, o algún item sin text
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: No es admin
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /items/{id}:
    put:
      tags: [items]
      summary: Reemplazar un item
      operationId: updateItem
      description: Requiere rol admin. Re-embebe el texto y reemplaza los vectores.
      parameters:
        - $ref: "#/components/parameters/ItemId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InsertRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InsertResponse" }
        "404":
          description: No existe
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: No es admin
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    delete:
      tags: [items]
      summary: Borrar un item
      operationId: deleteItem
      description: Requiere rol admin.
      parameters:
        - $ref: "#/components/parameters/ItemId"
      responses:
        "204":
          description: Borrado
        "404":
          description: No existe
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: No es admin
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /search:
    get:
      tags: [items]
      summary: Búsqueda semántica en vec_items
      operationId: searchItems
      description: Cualquier usuario autenticado.
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
          description: Texto a buscar (se embebe y se compara por distancia).
        - name: limit
          in: query
          schema: { type: integer, default: 10, minimum: 1 }
        - name: rerank
          in: query
          schema: { type: boolean, default: true }
          description: "true = float32 exacto (default); false = binario, ~20x más rápido, ~5% menos preciso."
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SearchResult" }
        "400":
          description: Falta q
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /collections:
    post:
      tags: [collections]
      summary: Crear una colección
      operationId: createCollection
      description: Requiere rol admin. dimensions es obligatorio (1..8192) cuando vector es true. Toda colección referenciada debe existir ya.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCollectionRequest" }
      responses:
        "201":
          description: Creada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Collection" }
        "400":
          description: Nombre inválido, dimensions inválido, o referencia a colección inexistente
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: No es admin
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: Ya existe una colección con ese nombre
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    get:
      tags: [collections]
      summary: Listar colecciones
      operationId: listCollections
      description: Cualquier usuario autenticado.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Collection" }
  /collections/{name}:
    delete:
      tags: [collections]
      summary: Borrar una colección entera
      operationId: dropCollection
      description: Requiere rol admin.
      parameters:
        - $ref: "#/components/parameters/CollectionName"
      responses:
        "204":
          description: Borrada
        "404":
          description: No existe
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: No es admin
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /collections/{name}/rename:
    put:
      tags: [collections]
      summary: Cambiar el nombre de una colección
      operationId: renameCollection
      description: >
        Requiere rol admin. Solo cambia el nombre lógico — items, vectores,
        índice full-text y referencias (propias y entrantes de otras
        colecciones) quedan intactos, actualizados al nuevo nombre.
      parameters:
        - $ref: "#/components/parameters/CollectionName"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RenameCollectionRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Collection" }
        "400":
          description: Falta name, o el nombre nuevo es inválido
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: No es admin
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Colección no existe
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: Ya existe una colección con el nombre nuevo
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /collections/{name}/items:
    post:
      tags: [collections]
      summary: Insertar un item en una colección
      operationId: insertDocument
      description: Requiere rol admin. text es obligatorio solo si la colección tiene vector.
      parameters:
        - $ref: "#/components/parameters/CollectionName"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DocumentRequest" }
      responses:
        "201":
          description: Creado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InsertResponse" }
        "400":
          description: Falta text en colección con vector, o referencia inválida
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: No es admin
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Colección no existe
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    get:
      tags: [collections]
      summary: Listar items de una colección
      operationId: listDocuments
      description: >
        Cualquier usuario autenticado. Cualquier query param no reservado
        filtra por un campo de nivel superior de data: campo=valor (equals)
        o campo__op=valor con op en eq, ne, lt, lte, gt, gte, like. Varios
        params se combinan con AND.
      parameters:
        - $ref: "#/components/parameters/CollectionName"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/Sort"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Document" }
        "404":
          description: Colección no existe
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /collections/{name}/items/bulk:
    post:
      tags: [collections]
      summary: Insertar varios items en una colección
      operationId: bulkInsertDocuments
      description: >
        Requiere rol admin. Mismo todo-o-nada que /items/bulk. Máximo 100
        items por request.
      parameters:
        - $ref: "#/components/parameters/CollectionName"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items: { $ref: "#/components/schemas/DocumentRequest" }
      responses:
        "201":
          description: Creados, mismo orden que el request
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/InsertResponse" }
        "400":
          description: Lote vacío, supera 100 items, o algún item sin text (colecciones con vector)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: No es admin
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Colección no existe
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /collections/{name}/items/{id}:
    get:
      tags: [collections]
      summary: Traer un item por id
      operationId: getDocument
      description: Cualquier usuario autenticado. Resuelve campos de referencia declarados en la colección.
      parameters:
        - $ref: "#/components/parameters/CollectionName"
        - $ref: "#/components/parameters/ItemId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Document" }
        "404":
          description: Colección o item no existe
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    put:
      tags: [collections]
      summary: Reemplazar un item
      operationId: updateDocument
      description: Requiere rol admin.
      parameters:
        - $ref: "#/components/parameters/CollectionName"
        - $ref: "#/components/parameters/ItemId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DocumentRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InsertResponse" }
        "400":
          description: Falta text en colección con vector, o referencia inválida
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: No es admin
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Colección o item no existe
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    delete:
      tags: [collections]
      summary: Borrar un item
      operationId: deleteDocument
      description: Requiere rol admin. Respeta referencias entrantes (restrict/set_null).
      parameters:
        - $ref: "#/components/parameters/CollectionName"
        - $ref: "#/components/parameters/ItemId"
      responses:
        "204":
          description: Borrado
        "403":
          description: No es admin
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Colección o item no existe
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: Bloqueado por una referencia restrict
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: >
        Token HMAC stateless obtenido de POST /login (no es JWT estándar,
        pero se usa igual: header Authorization: Bearer <token>).
  parameters:
...
    Username:
      name: username
      in: path
      required: true
      schema: { type: string }
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
      required: [error]
    BackupInfo:
      type: object
      properties:
        name: { type: string }
        size_bytes: { type: integer }
        created_at: { type: string, format: date-time }
      required: [name, size_bytes, created_at]
...
    LoginRequest:
      type: object
      properties:
        username: { type: string }
        password: { type: string }
      required: [username, password]
    LoginResponse:
      type: object
      properties:
        token: { type: string }
        expires_in: { type: integer, description: "Segundos hasta la expiración." }
      required: [token, expires_in]
    ChangePasswordRequest:
      type: object
      properties:
        current_password: { type: string }
        new_password: { type: string, minLength: 8 }
      required: [current_password, new_password]
    ResetPasswordRequest:
      type: object
      properties:
        new_password: { type: string, minLength: 8 }
      required: [new_password]
    CreateUserRequest:
      type: object
      properties:
        username:
          type: string
          pattern: "^[a-zA-Z][a-zA-Z0-9_]{0,62}$"
        password: { type: string, minLength: 8 }
        role: { type: string, enum: [admin, read-only] }
      required: [username, password, role]
    UserInfo:
      type: object
      properties:
        username: { type: string }
        role: { type: string, enum: [admin, read-only] }
      required: [username, role]
...